Revert name to reduce PR impact
