+++ anomdetect/src/main/java/com/expedia/adaptivealerting/anomdetect/forecast/point/PointForecasterParams.java
@@ -15,14 +15,14 @@
  */
 package com.expedia.adaptivealerting.anomdetect.forecast.point;
 
-import com.expedia.adaptivealerting.anomdetect.forecast.point.holtwinters.HoltWintersPointForecasterParams;
+import com.expedia.adaptivealerting.anomdetect.forecast.point.holtwinters.HoltWintersForecaster;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME)
 @JsonSubTypes({
         @JsonSubTypes.Type(value = EwmaPointForecaster.Params.class, name = "ewma"),
-        @JsonSubTypes.Type(value = HoltWintersPointForecasterParams.class, name = "holt-winters"),
+        @JsonSubTypes.Type(value = HoltWintersForecaster.Params.class, name = "holt-winters"),
         @JsonSubTypes.Type(value = PewmaPointForecaster.Params.class, name = "pewma"),
 })
 public interface PointForecasterParams {
