+++ anomdetect/src/main/java/com/expedia/adaptivealerting/anomdetect/forecast/point/holtwinters/HoltWintersOnlineAlgorithm.java
@@ -29,7 +29,7 @@ public class HoltWintersOnlineAlgorithm {
      * @param params     Contains the parameters for model
      * @param components Contains the online values used to calculate level, base and seasonality components (based on most recent history from observation at t-1).
      */
-    public void observeValueAndUpdateForecast(double y, HoltWintersPointForecasterParams params, HoltWintersOnlineComponents components) {
+    public void observeValueAndUpdateForecast(double y, HoltWintersForecaster.Params params, HoltWintersOnlineComponents components) {
         // Retrieve the model's parameters as set by user (or defaults)
         double alpha = params.getAlpha();
         double beta = params.getBeta();
