+++ anomdetect/src/main/java/com/expedia/adaptivealerting/anomdetect/forecast/point/holtwinters/HoltWintersSimpleTrainingModel.java
@@ -34,7 +34,7 @@ public class HoltWintersSimpleTrainingModel {
     private final double[] firstCycle;
     private final double[] secondCycle;
 
-    public HoltWintersSimpleTrainingModel(HoltWintersPointForecasterParams params) {
+    public HoltWintersSimpleTrainingModel(HoltWintersForecaster.Params params) {
         this.firstCycle = new double[params.getFrequency()];
         this.secondCycle = new double[params.getFrequency()];
     }
@@ -55,7 +55,7 @@ public class HoltWintersSimpleTrainingModel {
      * @param params     model parameters
      * @param components model components
      */
-    public void observeAndTrain(double y, HoltWintersPointForecasterParams params, HoltWintersOnlineComponents components) {
+    public void observeAndTrain(double y, HoltWintersForecaster.Params params, HoltWintersOnlineComponents components) {
         checkNulls(params, components);
         checkTrainingMethod(params);
         checkStillInInitialTraining(params);
@@ -77,14 +77,14 @@ public class HoltWintersSimpleTrainingModel {
         n++;
     }
 
-    public boolean isTrainingComplete(HoltWintersPointForecasterParams params) {
+    public boolean isTrainingComplete(HoltWintersForecaster.Params params) {
         return n >= (params.calculateInitTrainingPeriod());
     }
 
     /**
      * Update the level, base and seasonal components by running the main algorithm over each of the observations to this point.
      */
-    private void updateComponentsAndForecast(HoltWintersPointForecasterParams params, HoltWintersOnlineComponents components) {
+    private void updateComponentsAndForecast(HoltWintersForecaster.Params params, HoltWintersOnlineComponents components) {
         HoltWintersOnlineAlgorithm algorithm = new HoltWintersOnlineAlgorithm();
         concat(Arrays.stream(firstCycle), Arrays.stream(secondCycle)).forEach(y -> algorithm.observeValueAndUpdateForecast(y, params, components));
     }
@@ -93,12 +93,12 @@ public class HoltWintersSimpleTrainingModel {
         components.setLevel(mean(firstCycle));
     }
 
-    private void setBase(HoltWintersPointForecasterParams params, HoltWintersOnlineComponents components) {
+    private void setBase(HoltWintersForecaster.Params params, HoltWintersOnlineComponents components) {
         double base = (mean(secondCycle) - components.getLevel()) / params.getFrequency();
         components.setBase(base);
     }
 
-    private void setSeasonals(double y, HoltWintersPointForecasterParams params, HoltWintersOnlineComponents components) {
+    private void setSeasonals(double y, HoltWintersForecaster.Params params, HoltWintersOnlineComponents components) {
         for (int i = 0; i < params.getFrequency(); i++) {
             double s = params.isMultiplicative()
                     ? firstCycle[i] / components.getLevel()
@@ -107,17 +107,17 @@ public class HoltWintersSimpleTrainingModel {
         }
     }
 
-    private void checkNulls(HoltWintersPointForecasterParams params, HoltWintersOnlineComponents components) {
+    private void checkNulls(HoltWintersForecaster.Params params, HoltWintersOnlineComponents components) {
         notNull(params, "params can't be null");
         notNull(components, "components can't be null");
     }
 
-    private void checkTrainingMethod(HoltWintersPointForecasterParams params) {
+    private void checkTrainingMethod(HoltWintersForecaster.Params params) {
         isTrue(HoltWintersTrainingMethod.SIMPLE.equals(params.getInitTrainingMethod()),
                 String.format("Expected training method to be %s but was %s", HoltWintersTrainingMethod.SIMPLE, params.getInitTrainingMethod()));
     }
 
-    private void checkStillInInitialTraining(HoltWintersPointForecasterParams params) {
+    private void checkStillInInitialTraining(HoltWintersForecaster.Params params) {
         isFalse(isTrainingComplete(params),
                 String.format("Training invoked %d times which is greater than the training window of frequency * 2 (%d * 2 = %d) observations.",
                         n + 1, params.getFrequency(), params.calculateInitTrainingPeriod()));
