+++ anomdetect/src/test/java/com/expedia/adaptivealerting/anomdetect/forecast/point/holtwinters/HoltWintersSimpleTrainingModelTest.java
@@ -42,7 +42,7 @@ public class HoltWintersSimpleTrainingModelTest {
     public void testNullComponentsFails() {
         expectedEx.expect(IllegalArgumentException.class);
         expectedEx.expectMessage("components can't be null");
-        HoltWintersPointForecasterParams params = HoltWintersAustouristsTestHelper.buildAustouristsParams(SeasonalityType.MULTIPLICATIVE);
+        HoltWintersForecaster.Params params = HoltWintersAustouristsTestHelper.buildAustouristsParams(SeasonalityType.MULTIPLICATIVE);
         HoltWintersSimpleTrainingModel subject = new HoltWintersSimpleTrainingModel(params);
         subject.observeAndTrain(0, params, null);
     }
@@ -51,7 +51,7 @@ public class HoltWintersSimpleTrainingModelTest {
     public void testInvalidTrainingMethod() {
         expectedEx.expect(IllegalArgumentException.class);
         expectedEx.expectMessage(String.format("Expected training method to be %s but was %s", HoltWintersTrainingMethod.SIMPLE, HoltWintersTrainingMethod.NONE));
-        HoltWintersPointForecasterParams params = HoltWintersAustouristsTestHelper.buildAustouristsParams(SeasonalityType.MULTIPLICATIVE);
+        HoltWintersForecaster.Params params = HoltWintersAustouristsTestHelper.buildAustouristsParams(SeasonalityType.MULTIPLICATIVE);
         HoltWintersOnlineComponents components = new HoltWintersOnlineComponents(params);
         HoltWintersSimpleTrainingModel subject = new HoltWintersSimpleTrainingModel(params);
         subject.observeAndTrain(0, params, components);
@@ -63,7 +63,7 @@ public class HoltWintersSimpleTrainingModelTest {
         expectedEx.expectMessage(String.format(
                 "Training invoked %d times which is greater than the training window of frequency * 2 (%d * 2 = %d) observations.",
                 (HoltWintersAustouristsTestHelper.AUSTOURISTS_FREQUENCY * 2) + 1, HoltWintersAustouristsTestHelper.AUSTOURISTS_FREQUENCY, HoltWintersAustouristsTestHelper.AUSTOURISTS_FREQUENCY * 2));
-        HoltWintersPointForecasterParams params = HoltWintersAustouristsTestHelper.buildAustouristsParams(SeasonalityType.MULTIPLICATIVE)
+        HoltWintersForecaster.Params params = HoltWintersAustouristsTestHelper.buildAustouristsParams(SeasonalityType.MULTIPLICATIVE)
                 .setInitTrainingMethod(HoltWintersTrainingMethod.SIMPLE);
         HoltWintersOnlineComponents components = new HoltWintersOnlineComponents(params);
         HoltWintersSimpleTrainingModel subject = new HoltWintersSimpleTrainingModel(params);
@@ -73,7 +73,7 @@ public class HoltWintersSimpleTrainingModelTest {
     }
 
     private void checkObserveAndTrain(SeasonalityType seasonalityType, double expectedLevel, double expectedBase, double[] expectedSeasonal) {
-        HoltWintersPointForecasterParams params = HoltWintersAustouristsTestHelper.buildAustouristsParams(seasonalityType)
+        HoltWintersForecaster.Params params = HoltWintersAustouristsTestHelper.buildAustouristsParams(seasonalityType)
                 .setInitTrainingMethod(HoltWintersTrainingMethod.SIMPLE);
         HoltWintersOnlineComponents components = new HoltWintersOnlineComponents(params);
         HoltWintersSimpleTrainingModel subject = new HoltWintersSimpleTrainingModel(params);
